Доработки перед курсовой
