+++ Oborydovanie_Client/Oborydovanie_Client.csproj
@@ -134,6 +134,7 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>ModelOborudovanie.edmx</DependentUpon>
     </Compile>
+    <Compile Include="DataBase\NewClient.cs" />
     <Compile Include="DataBase\Order.cs">
       <DependentUpon>ModelOborudovanie.tt</DependentUpon>
     </Compile>
