+++ Oborydovanue_Manager/Pages/DobavSupplier.xaml.cs
@@ -48,7 +48,7 @@ namespace Oborydovanue_Manager.Pages
         }
         private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
         {
-            OrganizationPAgeWork.Instance.IsEnabled = true;
+            ManagerWork.Instance.IsEnabled = true;
         }
     }
 }
