+++ Oborydovanue_StoreAdmin/Style/ListStl.xaml
@@ -1,32 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <!--<Style x:Key="ListViewStyle" TargetType="{x:Type ListView}">
-        <Setter Property="Foreground" Value="White"/>
-        <Setter Property="Background" Value="Transparent"/>
-        <Setter Property="BorderBrush" Value="Transparent"/>
-        <Setter Property="BorderThickness" Value="0"/>
-        <Setter Property="Margin" Value="5"/>
-        <Setter Property="Padding" Value="0"/>
-        <Setter Property="FontWeight" Value="Normal"/>
-        <Setter Property="FontFamily" Value="Segoe UI"/>
-        <Setter Property="FontSize" Value="18"/>
-        <Setter Property="Template">
-            <Setter.Value>
-                <ControlTemplate TargetType="{x:Type ListView}">
-                    <Border Background="{TemplateBinding Background}"
-                            BorderBrush="{TemplateBinding BorderBrush}"
-                            BorderThickness="{TemplateBinding BorderThickness}"
-                            CornerRadius="10">
-                       
-                        <ScrollViewer Background="Transparent">
-                            <ItemsPresenter/>
-                        </ScrollViewer>
-                    </Border>
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
-    </Style>-->
-    
+
     <Style x:Key="ListViewStyle" TargetType="{x:Type ListView}">
         <Setter Property="Foreground" Value="White"/>
         <Setter Property="Background" Value="#333333"/>
@@ -51,42 +25,5 @@
                 </ControlTemplate>
             </Setter.Value>
         </Setter>
-
-        <!--<Setter Property="ItemContainerStyle">
-            <Setter.Value>
-                <Style TargetType="{x:Type ListViewItem}">
-
-                    <Setter Property="Background" Value="Transparent"/>
-
-                    <Setter Property="Template">
-                        <Setter.Value>
-
-                            <ControlTemplate TargetType="{x:Type ListViewItem}">
-                                <Border Background="{TemplateBinding Background}"
-                                        Padding="5">
-                                    <ContentPresenter/>
-                                </Border>
-
-                                <ControlTemplate.Triggers>
-                                    <Trigger Property="IsMouseOver" Value="True">
-                                        <Setter Property="Background" Value="LightBlue"/>
-                                    </Trigger>
-
-                                    <Trigger Property="IsMouseCaptured" Value="True">
-                                        <Setter Property="Background" Value="MistyRose"/>
-                                    </Trigger>
-
-                                    <Trigger Property="IsSelected" Value="True">
-                                        <Setter Property="Background" Value="#BF1821B8"/>
-                                    </Trigger>
-                                </ControlTemplate.Triggers>
-                            </ControlTemplate>
-
-                        </Setter.Value>
-                    </Setter>
-                </Style>
-            </Setter.Value>
-        </Setter>-->
-       
     </Style>
 </ResourceDictionary>
